+++ packages/widget_toolkit_pin/lib/src/lib_pin_code_with_biometrics/blocs/pin_code_bloc.dart
@@ -49,7 +49,7 @@ class PinCodeBloc extends $PinCodeBloc {
     required this.biometricAuthenticationService,
     required this.pinCodeService,
     required this.localizedReason,
-    required this.autoBiometricAuth,
+    required this.autoPromptBiometric,
   }) {
     authenticated.connect().addTo(_compositeSubscription);
   }
@@ -57,7 +57,7 @@ class PinCodeBloc extends $PinCodeBloc {
   final PinBiometricsService biometricAuthenticationService;
   final PinCodeService pinCodeService;
   final String localizedReason;
-  final bool autoBiometricAuth;
+  final bool autoPromptBiometric;
 
   final BehaviorSubject<String> _pinCode = BehaviorSubject.seeded('');
 
@@ -92,15 +92,12 @@ class PinCodeBloc extends $PinCodeBloc {
         _digitsCountState.switchMap((digitsCount) =>
             _checkPin(_pinCode.value, digitsCount).asResultStream()),
         _$biometricsButtonPressedEvent
-            .switchMap((_) => _authenticateWithBiometrics().asResultStream()),
-        Rx.combineLatest2<bool, bool, bool>(
-          showBiometricsButton,
-          Stream.value(autoBiometricAuth),
-          (showBiometricsButton, autoBiometricAuth) =>
-              showBiometricsButton && autoBiometricAuth,
-        )
-            .where((shouldAuthenticate) => shouldAuthenticate)
-            .switchMap((_) => _authenticateWithBiometrics().asResultStream()),
+            .mapTo(true)
+            .startWith(autoPromptBiometric)
+            .where((shouldPrompt) => shouldPrompt)
+            .asyncMap((_) => _getAreBiometricsEnabled())
+            .where((biometricsEnabled) => biometricsEnabled)
+            .switchMap((_) => _authenticateWithBiometrics().asResultStream())
       ]).setResultStateHandler(this).whereSuccess().publish();
 
   @override