+++ packages/widget_toolkit_pin/lib/src/lib_pin_code_with_biometrics/di/pin_code_dependencies.dart
@@ -14,7 +14,7 @@ class PinCodeDependencies {
     this.biometricsLocalDataSource,
     this.localizedReason,
     this.biometricsAuthDataSource,
-    this.autoBiometricAuth,
+    this.autoPromptBiometric,
   );
 
   factory PinCodeDependencies.from({
@@ -22,21 +22,21 @@ class PinCodeDependencies {
     required BiometricsLocalDataSource? biometricsLocalDataSource,
     required String localizedReason,
     PinBiometricsAuthDataSource? biometricsAuthDataSource,
-    bool autoBiometricAuth = false,
+    bool autoPromptBiometric = false,
   }) =>
       PinCodeDependencies._(
         pinCodeService,
         biometricsLocalDataSource,
         localizedReason,
         biometricsAuthDataSource,
-        autoBiometricAuth,
+        autoPromptBiometric,
       );
 
   final String localizedReason;
   final PinCodeService pinCodeService;
   final BiometricsLocalDataSource? biometricsLocalDataSource;
   final PinBiometricsAuthDataSource? biometricsAuthDataSource;
-  final bool autoBiometricAuth;
+  final bool autoPromptBiometric;
 
   late List<SingleChildWidget> providers = [
     ..._localAuthentication,
@@ -92,7 +92,7 @@ class PinCodeDependencies {
         biometricAuthenticationService: context.read<PinBiometricsService>(),
         pinCodeService: pinCodeService,
         localizedReason: localizedReason,
-        autoBiometricAuth: autoBiometricAuth,
+        autoPromptBiometric: autoPromptBiometric,
       ),
     ),
   ];