+++ packages/widget_toolkit_pin/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [0.3.0]
-* Added `autoBiometricAuth` parameter to `PinCodeKeyboard` to automatically authenticate with biometrics if available
+* Added `autoPromptBiometric` parameter to `PinCodeKeyboard` to automatically authenticate with biometrics if available
 * Fixed a bug where biometrics authentication would not work if encryption was used
 * Added biometrics authentication to the example application